Label logic and new unit tests
